Make directly includeable from source
Move all header files to a subdirectory named identical to the directory
where the libraries will be installed: this allows other programs to
compile against the libmaia source directory instead of having to do an
install.

For CMake, also add the current source directory to the list of
include-directories in the interface.

For QMake, also add two more library files to be installed.