MaiaXmlRpcServer: add local copy of allowedAddresses
It is often desirable to be able to set a local copy of some property
instead of having to manually keep track of which object holds observing
pointers to which other variable. To this end, add a
setAllowedAddresses() overload that takes a local copy of the address
list. Keep the original pattern for compatibility.