MaiaXmlRpcServer: allow non-listening construction
It is customary to be able to construct a server without it already
being listing. This also simplifies checking for errors when starting to
listen.

Note: for backwards compatibility, keep the constructor without any
arguments as the one that listens to any interface on port 8080 right
after construction. This is certainly not how other server objects
behave, but this would silently break existing code otherwise.